Added delta time scroll

+++ Source/Guis/GuiHorizontalBox.h
@@ -34,7 +34,7 @@ namespace ClassicLauncher
         GuiMiniCover* mMiniCover;
         std::vector<GuiCard*> mGuiCards;
         int mCardPositions[10]{ -632 - 6, -376 - 6, -120 - 6, 136 - 6, 392 - 6, 648 - 6, 904 - 6, 1160 - 6, 1416 - 6, 1672 - 6 };
-        int mPositionX;
+        float mPositionX;
         bool mIsLeft;
         bool mIsRight;
         Direction mLastDirection;
